Modified LT_INIT order, fix .la file
-In order for the generated .la file to make any sense, LT_INIT needs to
be run after some important CFLAGS are set. While I don't like setting
CFLAGS and CXXFLAGS directly in configure.ac, the only flags that are
set there are flags that are added due to user input to the configure
script. The important flags have to do with floating point ABI and
architecture, without which the wrong libstdc++.a is included in the
generated .la file.